requirements.txt
anyio==4.8.0
apsw==3.49.1.0
apswutils==0.0.2
beautifulsoup4==4.13.3
certifi==2025.1.31
click==8.1.8
fastcore==1.7.29
fastlite==0.1.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.12.4
python-multipart==0.0.20
pyyaml==6.0.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.0
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
